Added options for automatically reading the filters from HDF5 file.
The ``config.py`` file has a new option for *filter_1_material* and
*filter_2_material*. The value 'auto' will instruct tomopy-cli to read
the filter configuration from the measurement meta-data stored in the
HDF5 file. If the filter is open, no meta-data are found, or the
meta-data are not in the expected format, a 0 µm thick Al filter will
be used.

This commit also cleans up some of the logging calls for indentation,
etc., and fixes a bug in automatic rotation finding.