Files listed with rotation centers in a json file can now include subdirectories.
This fixes an issue where files can be listed in a rotation_axis.json
file but cannot be found because the parent directory is not included.

The algorithm starts at the file-system root and tries to match any
relative path up to the individual file name.

Example: *my_project/rotation_axis.json* points to files in
subdirectories *my_project/subprojectA/tomogram.h5* and
*my_project/subprojectB/tomogram.h5*. There would be a collision when
trying to get the rotation center because both files would be listed
as ``tomogram.h5`` in the json file.

With this change, the json file can now list "subprojectA/tomogram.h5"
and "subprojectB/tomogram.h5". If "tomogram.h5" is also listed
separately, the entries for "subprojectA/tomogram.h5" and
"subprojectB/tomogram.h5" would take precendence if they match the
argument to ``--file-name``.