Removed the "yaml" option for ``--rotation-axis-auto`` parameter.
This option would direct the rotation_axis to be taken from the YAML
file, to match the behavior of the old
``--rotation-axis-auto=json``. But with the new YAML
``--parameter-file`` option, this is redundant since
``--rotation-axis-auto=manual`` with do this as well. If a specific
rotation axis should be used to override the YAML file, this can be
given as a command line argument.