Refactored beam hardening code.
Previously, a lot of intermediate and state variables were being used
as global, module-level variables. I suspect this is the cause of a
bug that produces improper beam hardening corrections when working on
a large set of tomograms in one session.

This refactoring captures all the variables in a *BeamSoftener* class,
which is created fresh with each run through the tomopy-cli execution
path (in prep.py). This does mean source data has to be reloaded every
time. If this turns out to be a problem, we can fix it then
("premature opitimization is the root of all evil").